move math_utils under utils subpackage
