+++ src/funman/utils/__init__.py
@@ -1,0 +1,3 @@
+from . import math_utils
+
+__all__ = ["math_utils"]
