+++ src/funman/utils/math_utils.py
